Move SetupTokenInfo out of CardPayments module.
