+++ Demo/src/main/java/com/paypal/android/ui/vaultcard/VaultCardUiState.kt
@@ -2,7 +2,7 @@ package com.paypal.android.ui.vaultcard
 
 import com.paypal.android.api.model.CardPaymentToken
 import com.paypal.android.api.model.CardSetupToken
-import com.paypal.android.cardpayments.SetupTokenInfo
+import com.paypal.android.ui.approveorder.SetupTokenInfo
 import com.paypal.android.cardpayments.threedsecure.SCA
 import com.paypal.android.uishared.state.ActionState
