+++ Demo/src/main/java/com/paypal/android/ui/vaultcard/VaultCardViewModel.kt
@@ -13,7 +13,7 @@ import com.paypal.android.cardpayments.CardFinishVaultResult
 import com.paypal.android.cardpayments.CardPresentAuthChallengeResult
 import com.paypal.android.cardpayments.CardVaultRequest
 import com.paypal.android.cardpayments.CardVaultResult
-import com.paypal.android.cardpayments.SetupTokenInfo
+import com.paypal.android.ui.approveorder.SetupTokenInfo
 import com.paypal.android.cardpayments.threedsecure.SCA
 import com.paypal.android.corepayments.CoreConfig
 import com.paypal.android.models.TestCard
