+++ Demo/src/main/java/com/paypal/android/uishared/components/CardVaultResultView.kt
@@ -4,7 +4,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.paypal.android.cardpayments.SetupTokenInfo
+import com.paypal.android.ui.approveorder.SetupTokenInfo
 import com.paypal.android.utils.UIConstants
 
 @Composable
